feat: refactor clientjs to support err handling and notifications
Clientjs needed a refactor, because error states from
failed transports and notifications would silently
fail or would require exceptional handling. Additionally
the refactor adds support for timeout, and eventual
implementation of request strategies for choosing
which transport to use in a request.

This refactor exposes notifications handler, and error handler
as well as a timeout handler for timing out request. The refactor
also moves request resolution to be handle by the transport,
and transport selection and batching start and stopping to be
handled by the request manager freeing multiple transports
to potentially be used at the same time.